Reconcile(?) cube coords and cubie coords
Both should now be standardized on (front-ness, down-ness, right-ness).
This means that numpy's pretty-print order roughly matches the layout of
the cube (the second axis is arranged vertically and the third
horizontally; the first controls the page so the correspondence is a
little less direct), but the coordinate system is still right-handed.
(My intuition would be for the pages to go front-to-back, but that's a
left-handed coordinate system, so I've instead declared that they go
back-to-front.)